refactor(desktop): share managed runtime signature checks
